feat: The API Rest of the application has been created.
The following files were added: usersController, users and usersRoutes. The usersControllers is dedicated to be the mediator between the server database and the requests to it. usersRoutes is the one who has the route of the request to call the controller. Finally, the users file is the model that contains the table data.