feat: Role assigned each time a user is created.
Also, when a user is created, the server returns in the body the data of the user. This is important because the app save this info in a SharedPreferences. The table "user_has_roles" is the manager to organice each user with the id role, where: 1 is Client, 2 is restaurant and 3 is delivery.