feat: User can select his profile picture.
This feature was make it with a library called ImagePicker and it's very cool. You can choose a picture with your camera or the intern files in your phone. The unique problem about this library, it is when you select a picture on your files because the application don´t request you the permission.