feat: User can update his profile info.
There is a problem. When the user update only parameters like name, last name or phone, the application turn crash.

The reason it's because the image file to send to the server is null. So, it's necessary change the photo to update the user. This is a problem.

Next commit the solution would be implemented a new route in the server. Update the user info without image, for example.

Problem: #160 in ClientUpdateActivity.kt.
Reason: NullPointerException.