feat: Client can see categories in the home fragment.
I had a problem at the beginning with my phone. When I tried to make the request to the server, my phone failed to connect to the server.

I don't know the reason, but it was a problem with my phone, not the server or the code. The server was responding just fine with Postman.

What I did was to reboot my phone and disconnect it from the computer to get closer to the modem. Then it worked. Very rare.