feat: The product can be created with three images in the app.
There were minor changes in product.js. I forget to put some commas in the methods.

About Kotlin, the methodology it's the same with the other models. In this case, you need to be careful with send three images at the same time.

For do this, it's necessary send an arrayList of Files.