feat: Client can search products by category.
When the user clicks in a category, this navigates to product list activity and display he products in this category.

Each product is created by the restaurante and has like parameters: id, name, description, price, three images and quantity.

A minor change was make in productController.js. Before, the response returns a JSON object, but now, returns only the data.